+++ src/ChuckNorrisJokesServiceProvider.php
@@ -3,11 +3,17 @@
 namespace Nikulinilya\ChuckNorrisJokes;
 
 use Illuminate\Support\ServiceProvider;
+use Nikulinilya\ChuckNorrisJokes\Console\ChuckNorrisJoke;
 
 class ChuckNorrisJokesServiceProvider extends ServiceProvider
 {
     public function boot()
     {
+        if ($this->app->runningInConsole()) {
+            $this->commands([
+                ChuckNorrisJoke::class
+            ]);
+        }
     }
 
     public function register()