+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - ChuckNorrisJokesServiceProvider
 - Facades/ChuckNorris
 - laravel extra to composer.json
+- Console/ChuckNorrisJoke command
 
 ## [1.1.0] - 2020-05-24
 ### Added